re structure project and add window bar function
