--- index.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <meta http-equiv="Content-Security-Policy" content="default-src
-    'self'; script-src 'self'">
-    <link rel="stylesheet" type="text/css" href="semantic/dist/semantic.min.css">
-    <link rel="stylesheet" type="text/css" href="./style.css">
-    <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.3/dist/jquery.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/fomantic-ui@2.9.2/dist/semantic.min.css">
-    <script src="https://cdn.jsdelivr.net/npm/fomantic-ui@2.9.2/dist/semantic.min.js"></script>
-    <title>ElectraSnap</title>
-  </head>
-  <body>
-
-   <div class="container">
-    <img src="./icon-app.png" alt="">
-    <button class="ui teal button container-items">
-      <i class="instagram icon"></i>
-      Snapshot
-    </button>
-    <button class="ui brown button container-items">
-      <i class="history icon"></i>
-      Historys test
-    </button>
-   </div>
-   
-    <script src="./renderer.js"></script>
-  </body>
-</html>
