+++ src/preload.js
@@ -1,0 +1,6 @@
+const { contextBridge, ipcRenderer } = require('electron')
+
+contextBridge.exposeInMainWorld('electronAPI', {
+  mimizedWindow: () => ipcRenderer.send('mimizedWindow'),
+  appExit: () => ipcRenderer.send('appExit'),
+})
