+++ src/styles/style.css
@@ -1,0 +1,52 @@
+body {      
+    -webkit-app-region: drag;
+    background-color: transparent;
+}
+
+.container{
+    border-radius: 18px;
+    width: 100%;
+    height: 100vh;
+    background-color: #333;
+    color: white; 
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    gap: 8px;
+    
+}
+
+.container-item{
+    width: 40%;
+}
+
+.container-windowbar{
+    position: absolute;
+    right: 0;
+    top: 0;
+    margin: 8px;
+    -webkit-app-region: no-drag;
+}
+
+.container-windowbar-btn-minimize{
+    position: absolute;
+    top: -2px;
+    right: 30px;    
+    -webkit-app-region: no-drag;
+    
+} 
+
+.container-windowbar-btn-minimize:hover{
+    cursor: pointer;
+    color: #ffffffae;    
+    -webkit-app-region: no-drag;
+}
+
+.container-windowbar-btn-close:hover{
+    cursor: pointer;
+}
+
+.container-windowbar-btn-close{
+    color: #da5016; 
+}
