--- style.css
@@ -1,22 +1,0 @@
-body {      
-    -webkit-app-region: drag;
-    background-color: transparent;
-}
-
-.container{
-    border-radius: 18px;
-    width: 100%;
-    height: 100vh;
-    background-color: #333;
-    color: white; 
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    gap: 8px;
-    
-}
-
-.container-items{
-    width: 40%;
-}
